PNGs: save sRGB chunks (and appropriate cHRM/gAMA)
My current thinking is to save what we assume, and assume that the user
has calibrated their display to something approximating sRGB when working
with MyPaint.

Yes, this is a bad assumption, yes we should be using real display space
transformations *if* they are available: mypaint/mypaint#9. However it's a
backwards-compatible assumption for MyPaint's purposes, and hopefully a
minimally interoperable one.

Addresses one of the issues noticed in mypaint/mypaint#129.

The existence of non-sRGB-tagged PNG data in a .ORA file *cannot* be relied
upon to identify files saved with old stable MyPaint releases because the
GIMP ORA plugin has been known to save a gAMA chunk only (i.e. no cHRM, no
sRGB), housing a value that is (presumably) either derived from display
gamma or an sRGB assumed default.