Half-arsed separate painting thread
This is very incomplete. Looks like we'll need some careful treatment of
the tiledict when the brush engine needs to create new tiles, and it would
be good to release the GIL while doing pixel crunching.

Some UI update stuff invoked by the C++ stroke_to() needs to be bounced to
the main thread too: the way to do that is via an idle_add() call in the
subthread.